requirements.txt
# GPU-enabled PyTorch (CUDA 12.4 build — compatible with drivers up to CUDA 13.x)--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.2.0
torchvision
torchaudio
sentence-transformers>=2.7.0
transformers>=4.40.0

# Vector DB
chromadb>=0.5.0

# Clustering
scikit-learn>=1.4.0
scikit-fuzzy>=0.4.2
umap-learn>=0.5.6
numpy>=1.26.0
scipy>=1.13.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
notebook>=7.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0